Integrate pre-existing neural network code
