+++ EquationSolver/utilities/image/display.py
@@ -1,0 +1,7 @@
+import matplotlib.pyplot as plt
+
+
+def show(image):
+    plt.imshow(image)
+    plt.gray()
+    plt.show()
