+++ EquationSolver/utilities/neural_network/generator.py
@@ -1,23 +1,14 @@
-from argparse import ArgumentParser
-
 import numpy as np
 import sklearn.model_selection
 from tensorflow import keras
 
 
-def main():
-    parser = ArgumentParser(description="Generate neural network trained on MNIST dataset")
-    parser.add_argument("--benchmark", action="store_true", help="Benchmark the neural network using k-fold cross-validation")
-    parser.add_argument("--folds", type=int, default=5, metavar="", help="Set the number of folds for k-fold cross-validation")
-    args = parser.parse_args()
-
+def generate_neural_network(benchmark=False, number_of_folds=5):
     x, y = load_dataset()
-    x = normalize_pixels(x)
-
     model = generate_model()
 
-    if args.benchmark:
-        scores = evaluate_model(model, x, y, args.folds)
+    if benchmark:
+        scores = evaluate_model(model, x, y, number_of_folds)
         print("Accuracy: mean=%.3f std=%.3f, n=%d" % (np.mean(scores) * 100, np.std(scores) * 100, len(scores)))
     else:
         model.fit(x, y, epochs=5)
@@ -37,13 +28,6 @@ def load_dataset():
     y = np.concatenate((train_y, test_y))
 
     return x, y
-
-
-def normalize_pixels(train):
-    train_normalized = train.astype("float32")
-    train_normalized /= 255.0
-
-    return train_normalized
 
 
 def generate_model():
@@ -71,7 +55,3 @@ def evaluate_model(model, x, y, number_of_folds):
         accuracy = model.evaluate(test_x, test_y, verbose=0)[1]
         scores.append(accuracy)
     return scores
-
-
-if __name__ == "__main__":
-    main()
